Change loop variable types to size_t
Updated loop variable types from int to size_t for better compatibility with string length.

+++ aoc2025/day2/star2.cpp
@@ -6,20 +6,20 @@ using ll = long long;
 
 bool is_invalid(string s)
 {
-    for (int len = s.length() / 2; len >= 1; --len)
+    for (size_t len = s.length() / 2; len >= 1; --len)
     {
         if (s.length() % len != 0)
         {
             continue;
         }
         bool repeat{true};
-        for (int i = 0; i < len; ++i)
+        for (size_t i = 0; i < len; ++i)
         {
             if (!repeat)
             {
                 break;
             }
-            for (int j = i + len; j < s.length(); j += len)
+            for (size_t j = i + len; j < s.length(); j += len)
             {
                 if (s[j] != s[i])
                 {
